reveal works!!!

+++ src/State.ts
@@ -1,7 +1,7 @@
 import { TAbstractFile, TFile, Vault } from "obsidian";
-import { IdAllocator, parseDocument } from "./markdown";
+import { IdAllocator, parseDocument, SectionIdInfo } from "./markdown";
 import * as Doc from "./document";
-import Client from "./Client";
+import Client, { RevealMessage } from "./Client";
 
 function isMarkdownFile(maybeFile: TAbstractFile): TFile | null {
     if (!(maybeFile instanceof TFile) || maybeFile.extension != "md") {
@@ -10,8 +10,13 @@ function isMarkdownFile(maybeFile: TAbstractFile): TFile | null {
     return maybeFile;
 }
 
+export interface FrontendInterface {
+    goTo(file: string, line: number, column: number): void;
+}
+
 export class State {
-    documentByPath: Map<string, Doc.Root> = new Map();
+    documentByPath: Map<string, [Doc.Root, Doc.SectionId[]]> = new Map();
+    globalSectionIdMap: Map<Doc.SectionId, SectionIdInfo> = new Map();
     listeningClients: Client[] = [];
     /**
      * all section objects are given unique incrementing IDs.
@@ -19,31 +24,39 @@ export class State {
      */
     idAllocator: IdAllocator = new IdAllocator();
 
-    constructor(public vault: Vault) {}
+    constructor(public vault: Vault, public frontend: FrontendInterface) {}
 
 
     // Core index functions: get a path to a markdown file, and set or delete the contents
     indexSetByPathAndContents(path: string, contents: string): void {
-        const parsed = parseDocument(path, contents, this.idAllocator);
-        console.log("contents:", contents.slice(0, 50), "...");
-        console.log("parsed:", parsed);
-        if (parsed !== null) {
-            this.documentByPath.set(path, parsed);
+        const parseResult = parseDocument(path, contents, this.idAllocator);
+        if (parseResult !== null) {
+            const [doc, sectionIdMap] = parseResult;
+
+            const sectionIdList = Array.from(sectionIdMap.keys());
+            this.documentByPath.set(path, [doc, sectionIdList]);
+            sectionIdMap.forEach((v, k) => this.globalSectionIdMap.set(k, v));
+
             for (const c of this.listeningClients)
-                c.sendDocument(parsed);
+                c.sendDocument(doc);
         }
     }
 
     indexDeleteByPath(path: string): void {
-        this.documentByPath.delete(path);
-        for (const c of this.listeningClients)
-            c.removeDocument(path);
+        const entry = this.documentByPath.get(path);
+        if (entry !== undefined) {
+            const [_doc, sectionIdList] = entry;
+            this.documentByPath.delete(path);
+            sectionIdList.forEach(sectionId => this.globalSectionIdMap.delete(sectionId));
+            for (const c of this.listeningClients)
+                c.removeDocument(path);
+        }
     }
 
 
     addClient(c: Client) {
         // tell the client about all of the current documents
-        for (const document of this.documentByPath.values()) {
+        for (const [document, _sectionIdList] of this.documentByPath.values()) {
             c.sendDocument(document);
         }
         // this client listens to new changes
@@ -52,6 +65,27 @@ export class State {
 
     removeClient(c: Client) {
         this.listeningClients.remove(c);
+    }
+
+    // Client event handlers
+    onRevealMessage(msg: RevealMessage) {
+        const docId = msg.docId;
+        const entry = this.globalSectionIdMap.get(docId);
+        if (entry === undefined) {
+            return;
+        }
+
+        // this is an up to date entry and we have information about it
+        console.log(entry.mdNode);
+        const file = entry.filename;
+        const position = entry.mdNode.position;
+        if (position === undefined) {
+            return;
+        }
+
+        // go to file and position in the editor
+        // TODO: use the column itself rather than the start of the line. The column itself didn't act like I wanted it to.
+        this.frontend.goTo(file, position.start.line, 0);
     }
 
 
@@ -69,12 +103,6 @@ export class State {
     }
 
     vaultOnDelete(maybeFile: TAbstractFile): void {
-        // Only care about .md files
-        const file = isMarkdownFile(maybeFile);
-        if (file === null) {
-            return;
-        }
-
         this.indexDeleteByPath(maybeFile.path);
     }
